requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-taggit==1.5.1
django-tinymce==3.3.0
docopt==0.6.2
idna==3.3
Pillow==8.4.0
pipreqs==0.4.11
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
yarg==0.1.9
